Limpieza en el Backend

+++ CafeAppBackend/src/main/java/com/cafeapp/backend/controlador/UsuarioController.java
@@ -1,13 +1,9 @@
 package com.cafeapp.backend.controlador;
 
-import com.cafeapp.backend.dto.RegistroRequest;
-import com.cafeapp.backend.dto.UsuarioResponse;
-import com.cafeapp.backend.dto.LoginRequest;
 import com.cafeapp.backend.modelo.Usuario;
-import jakarta.validation.Valid;
+import com.cafeapp.backend.servicio.UsuarioService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import com.cafeapp.backend.servicio.UsuarioService;
 
 @RestController
 @RequestMapping("/api/usuarios")
@@ -19,21 +15,29 @@ public class UsuarioController {
         this.usuarioService = usuarioService;
     }
 
-
+    // ============================================================
+    // OBTENER POR ID
+    // ============================================================
     @GetMapping("/{id}")
-    public ResponseEntity<Usuario> obtener(@PathVariable Long id) {
+    public ResponseEntity<Usuario> obtener(@PathVariable Integer id) {
         Usuario usuario = usuarioService.obtenerPorId(id);
         return usuario != null ? ResponseEntity.ok(usuario) : ResponseEntity.notFound().build();
     }
 
+    // ============================================================
+    // ACTUALIZAR
+    // ============================================================
     @PutMapping("/{id}")
-    public ResponseEntity<Usuario> actualizar(@PathVariable Long id, @RequestBody Usuario usuario) {
+    public ResponseEntity<Usuario> actualizar(@PathVariable Integer id, @RequestBody Usuario usuario) {
         Usuario actualizado = usuarioService.actualizar(id, usuario);
         return actualizado != null ? ResponseEntity.ok(actualizado) : ResponseEntity.notFound().build();
     }
 
+    // ============================================================
+    // ELIMINAR
+    // ============================================================
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> eliminar(@PathVariable Long id) {
+    public ResponseEntity<Void> eliminar(@PathVariable Integer id) {
         boolean eliminado = usuarioService.eliminar(id);
         return eliminado ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
