+++ CafeAppBackend/src/main/java/com/cafeapp/backend/modelo/Curso.java
@@ -9,15 +9,14 @@ public class Curso {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_curso")
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String nombre;
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
 }
-
