+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/ItemCarritoRepository.java
@@ -3,5 +3,5 @@ package com.cafeapp.backend.repositorio;
 import com.cafeapp.backend.modelo.ItemCarrito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemCarritoRepository extends JpaRepository<ItemCarrito, Long> {
-}
+public interface ItemCarritoRepository extends JpaRepository<ItemCarrito, Integer> {}
+
