+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/RolRepository.java
@@ -4,6 +4,7 @@ package com.cafeapp.backend.repositorio;
 import com.cafeapp.backend.modelo.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Rol, Long> {
-}
+public interface RolRepository extends JpaRepository<Rol, Integer> {}
+
+
